Change region identifiers to lists of strings
Change region identifiers to lists of strings and create new type alias
RegionIdentifier. The region identifier was formerly a period seperated
list of region and subregion names (similar to the Haskell module
syntax, in addition to examples in other languages). This commit allows
region identifiers to easier to work with.