Changed radius of GraphView to a Word
Changed the type of the radius field of the GraphView type in the library
to a Word (a unsigned integer). This ensures the invariant that
the radius can never be negative, without use of smart constructors or
auxiliary update functions.

Remove the minimum and maximum radius fields in the GraphView type, as
they are no longer necessary.