Change the appearance of verticies to make them clearer
Reduce the length of the base of the triangle that represents an edge
to make overlapping edges clearer and give more visual space to the
vertices.

Change the vertex shader to render the darker interior of the circle
that represents a vertex by a change in color, not in opacity.

Render the edges first, so that the verticies will lie on top of them,
and therefore be more prominent and aesthetically pleasing.