Change the edge model
Change the edge model from a triangle to a trapezoid. Change the base
of the trapeziod from 0.2 to 0.25. Change the opposite end to have a
width of 0.05. The edge model tended to become invisible the closer it
got to the destination. These changes make the edge model more visible
at all distances.