Reduce code base to minimum rendering code
Remove code to the minimum required to render a window.

Remove the library/application distinction and simplify SConstruct
accordingly.

Move the `earlyExit` function declaration into a dedicated header,
`error.h`, to remove the necessity of including `stdio.h` to return
error messages. Modify `earlyExit` to accept and pass on variable
arguements to `vfprintf`.