Remove JGraphT in favor of trees
Remove JGraphT from project files and code to reduce distributable
size.

Rewrite internals to use a tree. `PersonalTrait` will now hold a
mutable list of references to it's children nodes.

Write two simple auxiliary functions for `PersonalTrait` to check if
one is a quality, and if a short description belongs to a child node.