Rewrite trait tree classes
Rewrite `PersonalTrait` to remove all references to qualities and
skills, as those are underdeveloped ideas.

Write a `PersonalTraitTree` class that stores a `HashMap` of `Int` to
`PersonalTrait`s. An indirect integer reference is better than directly
referencing a personal trait because the former is easier to serialize
and the map keeps all related personal traits together.

Change the node class, `Personal` trait, to reference the parent rather
than the children, as that reduces complexity and is valid for all
nodes. If a node has no parent, it's either a root or orphaned node.