Add trait change class
Add `PersonalTraitChange` class to keep track of changes to traits.

Write documentation for `PersonalTrait`.

Add initial, permanent nodes to `PersonalTraitTree`. The first node,
"Workspace," represents the workspace and is the root of the tree. The
second node, "Eudaemonia," represents the active subtree and is the
root of all non-orphaned nodes.