Position virtues in cartesian coordinates
Change the distance and angle members of the Virtue class, which
express a position in polar coordinates, to x and y members, which will
express positions in Cartesian coordinates. This change makes it easier
and natural to work with transformations and OpenGL, as Cartesian
coordinates are more common in introductory mathematics.