Rewrite UIComponent for encapsulation
Change UIComponent from an interface to a class. This change allows for
all methods to have a default implementation (id est, for all methods
that can be overrided by subclasses to have an implementation that does
nothing.)

Validate parameters to setDrawingArea and make it a public final
method. This change ensures all components will have predictable
behaviour.

Add a callback method, onSetDrawingArea, for subclasses to override in
order to update themselves after setDrawingArea has been called.