Revert "Split UIComponent"
This reverts commit 5f3e2452c3161027d17da1327581622ea38f7120.

The above commit was intended to allow a new window class and component
classes to share code. That change, however, would have allowed a
component to be initialized standalone with respect to a window, which
we do not want. Furthermore, this situation violates an important
principle of inheretance, which states that classes must be replacable
in functionality by their subclasses.