Add a SubLifecycle class
Add a class, SubLifecycle, with an internal lifecycle distinct from the
normal Java object lifecycle. This class will allow for the creation of
hierarchies of objects that require initialization from the top down
(see the reliance of components on the Window class).

This class still needs to be integrated into UIComponent and Window.