Rename the SubLifecycle class to Bounded
Rename the SubLifecycle class to Bounded, which better reflects that
objects of this class are objects bounded in time by an internal
lifecycle, rather than being the lifecycle of another object or being
their own lifecycle somehow.

Rename internal version of Bounded methods to begin with a underscore,
which better notes the similarities between the public final and
protected versions of these methods.