Rewrite the Viewport class
Make Viewport a immutable class. This change is generally considered
good design. Perhaps it should go into the contributing guidelines.

Create a private Viewport constructor for convenience, a public static
constructor for creating a Viewport from a GLFW window identifier, and
a public method constructor for creating a viewport that is within
another viewport.