Add Promises to AbstractView
Add a private member, viewportPromise, so that the view can be created
without requiring a Viewport.

Add a method, AbstractView.getViewport() to get the viewport. This
function will fall back to AbstractView.viewportPromise if
AbstractView.viewport is not set.

Remove AbstractView.getWidth() and AbstractView.getHeight(), as
subclasses can now retrieve the viewport themselves.