Simplify Viewport and AbstractView
Remove root tracking from Viewport. As it turn out, class
discrimination is very hard to implment with good Java and software
design practices. This method is useless, so this commit get's rid of
it and the overhead it brought.

Remove the promise contructor for AbstractView and instead repurpose
the AbstractView.setViewport() method.

Add a method to check if an AbstractMethod is initialized.

Remove promises.