Match packages to directory structure
Change the package of each Java source file to match it's directory.
Although Gradle appears to be fine with mismatching packages and
directories, this change allows for use to use the default
package-private access modifier for better safety. This change also
makes the source code make sense to potential contributors and
maintainers.