Launch Window from ViewRunnable
Launch a Window and run the main loop from ViewRunnable. All the
changes below are necessary to allow this to happen.

Enable a simple GLFWErrorCallback to facilitate debugging.

Fix GLFW minor OpenGL profile window hint.

Reorganize the main method to facilitate a simple change of threading
style.

Allow 0 for x- and y- offsets in view.Viewport. This should be allowed
because the window viewport and likely other viewports will have zero
offset relative to the bottom-left of the window.