Simplify component design
Join the presenter and view packages and threads to a single UI package
and thread, because the former is likely nearly as lightweight as the
latter and should have little effect on performance.

Remove the MVP model from contributing.md, as it was clearly an
inappropriate place to write about the overall program design.