Rename project to Virtuous Hummingbird
After much deliberation and anguish, we have decided to rename this
project to "Virtuous Hummingbird." The package name will change from
com.autumnara.aikaterine to com.autumnara.hummingbird. The current
name, Aikaterine, corresponds to a feminine Greek name in current use.
The use of feminine names in software design is suspicious and should
probably be avoided until the consequences are understood.